Added contexts and test steps for #19
