+++ tests/fixtures/web/accessibility/image-with-alt-text.html
@@ -2,8 +2,8 @@
 <html>
 <head>
     <meta charset="utf-8">
-    <title>Valid HTML</title>
-    <meta content="index, nofollow" name="robots"/>
+    <title>Alt Text</title>
+    <img src="http://localhost:8080/invent.jpg" alt="Design Thinking"/>
 </head>
 <body>
 <p>Valid HTML</p>
