+++ tests/fixtures/web/meta/robots/meta-robots-index-follow.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="utf-8">
     <title>Valid HTML</title>
-    <meta content="index, nofollow" name="robots"/>
+    <meta content="index, follow" name="robots"/>
 </head>
 <body>
 <p>Valid HTML</p>
