+++ tests/fixtures/web/meta/robots/meta-robots-noindex-follow.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="utf-8">
     <title>Valid HTML</title>
-    <meta content="noindex, nofollow" name="robots"/>
+    <meta content="noindex, follow" name="robots"/>
 </head>
 <body>
 <p>Valid HTML</p>
