+++ tests/fixtures/web/meta/robots/meta-robots-noindex-nofollow.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="utf-8">
     <title>Valid HTML</title>
-    <meta content="noindex, follow" name="robots"/>
+    <meta content="noindex, nofollow" name="robots"/>
 </head>
 <body>
 <p>Valid HTML</p>
