+++ tests/fixtures/web/ux/site-with-noviewport-tag.html
@@ -2,8 +2,7 @@
 <html>
 <head>
     <meta charset="utf-8">
-    <title>Valid HTML</title>
-    <meta content="index, nofollow" name="robots"/>
+    <title>Responsive Design</title>
 </head>
 <body>
 <p>Valid HTML</p>
