+++ tests/fixtures/web/ux/site-with-valid-viewport.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <meta charset="utf-8">
+    <title>Responsive Design</title>
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+</head>
+<body>
+<p>Valid HTML</p>
+</body>
+</html>
