+++ tests/fixtures/web/ux/site-without-valid-viewport.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <meta charset="utf-8">
+    <title>Responsive Design</title>
+    <meta name="viewport" content="initial-scale=1, maximum-scale=1, user-scalable=no">
+</head>
+<body>
+<p>Valid HTML</p>
+</body>
+</html>
